// client/src/DxDbe/src/models/Enums/InputType.ts
/*******************************************************************************
 * (C) Copyright
 * Logivations GmbH, Munich 2010-2022
 ******************************************************************************/

enum InputType {
	Autocomplete = 'dxAutocomplete',
	Calendar = 'dxCalendar',
	CheckBox = 'dxCheckBox',
	ColorBox = 'dxColorBox',
	DateBox = 'dxDateBox',
	DropDownBox = 'dxDropDownBox',
	HtmlEditor = 'dxHtmlEditor',
	Lookup = 'dxLookup',
	NumberBox = 'dxNumberBox',
	RadioGroup = 'dxRadioGroup',
	RangeSlider = 'dxRangeSlider',
	SelectBox = 'dxSelectBox',
	Slider = 'dxSlider',
	Switch = 'dxSwitch',
	TagBox = 'dxTagBox',
	TextArea = 'dxTextArea',
	TextBox = 'dxTextBox',
}

export default InputType;